refactor: rename config.py to context.py (#371)
